# app/api/openapi.yaml
openapi: 3.0.3
info:
    title: Foodies API
    version: '1.0.0'
    description: Foodies server API
    license:
        name: Apache 2.0
        url: https://www.apache.org/licenses/LICENSE-2.0
servers:
    - url: https://foodies-5tjw.onrender.com/api
      description: Foodies server
    - url: http://127.0.0.1:3000/api
      description: Local development server
...
paths:
...
    /users/register:
        post:
            security:
                - none: []
            tags:
                - users
            summary: Register user
            operationId: registerUser
            requestBody:
                $ref: '#/components/requestBodies/registerRequest'
            responses:
                '201':
                    $ref: '#/components/responses/registerResponse'
                '401':
                    $ref: '#/components/responses/notAuthorisedResponse'
                '500':
                    $ref: '#/components/responses/serverErrorResponse'
    /users/login:
        post:
            security:
                - none: []
            tags:
                - users
            summary: Login user
            operationId: loginUser
            requestBody:
                $ref: '#/components/requestBodies/loginRequest'
            responses:
                '200':
                    $ref: '#/components/responses/authorisedResponse'
                '401':
                    $ref: '#/components/responses/notAuthorisedResponse'
                '500':
                    $ref: '#/components/responses/serverErrorResponse'
...
    /users/logout:
        post:
            security:
                - bearerAuth: []
            tags:
                - users
            summary: Logout user
            operationId: logoutUser
            responses:
                '204':
                    description: No content
                '401':
                    $ref: '#/components/responses/notAuthorisedResponse'
                '500':
                    $ref: '#/components/responses/serverErrorResponse'
...
    /users/followees/{id}:
        post:
            security:
                - bearerAuth: []
            tags:
                - users
            summary: Follow a user
            operationId: follow
            parameters:
                - $ref: '#/components/parameters/id_path'
            responses:
                '204':
                    description: No content
                '400':
                    $ref: '#/components/responses/invalidRequestResponse'
                '401':
                    $ref: '#/components/responses/notAuthorisedResponse'
                '500':
                    $ref: '#/components/responses/serverErrorResponse'
        delete:
            security:
                - bearerAuth: []
            tags:
                - users
            summary: Unfollow a user
            operationId: unfollow
            parameters:
                - $ref: '#/components/parameters/id_path'
            responses:
                '204':
                    description: No content
                '401':
                    $ref: '#/components/responses/notAuthorisedResponse'
                '404':
                    $ref: '#/components/responses/notFoundResponse'
                '500':
                    $ref: '#/components/responses/serverErrorResponse'
    /users/{id}:
        get:
            security:
                - bearerAuth: []
            tags:
                - users
            summary: Get a user information
            operationId: getUserInformation
            parameters:
                - $ref: '#/components/parameters/id_path'
            responses:
                '200':
                    $ref: '#/components/responses/userResponse'
                '401':
                    $ref: '#/components/responses/notAuthorisedResponse'
                '404':
                    $ref: '#/components/responses/notFoundResponse'
                '500':
                    $ref: '#/components/responses/serverErrorResponse'
...
components:
    parameters:
        id_path:
            name: id
            in: path
            required: true
            schema:
                type: string
...
    requestBodies:
...
        registerRequest:
            description: Register user request
            required: true
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            name:
                                type: string
                            email:
                                type: string
                            password:
                                type: string
        loginRequest:
            description: Login user request
            required: true
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            email:
                                type: string
                            password:
                                type: string
    responses:
...
        userResponse:
            description: User information
            content:
                application/json:
                    schema:
                        allOf:
                            - $ref: '#/components/schemas/fullUser'
                            - type: object
                              properties:
                                  followers:
                                      type: integer
                                  followeers:
                                      type: integer
                                  recipes:
                                      type: integer
                                  favoriteRecipes:
                                      type: integer
        registerResponse:
            description: Successfully registered
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            user:
                                $ref: '#/components/schemas/fullUser'
        authorisedResponse:
            description: Successfully authorised
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            token:
                                type: string
                            user:
                                $ref: '#/components/schemas/fullUser'
...
        invalidRequestResponse:
            description: Invalid request
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/error'
        notFoundResponse:
            description: Not found
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/error'
        serverErrorResponse:
            description: Internal Server Error
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/error'
        notAuthorisedResponse:
            description: Not Authorised
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/error'
    schemas:
...
        fullUser:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                email:
                    type: string
                avatar:
                    type: string
...
        error:
            type: object
            properties:
                message:
                    type: string
    securitySchemes:
        none:
            type: http
            scheme: none
        bearerAuth: # arbitrary name for the security scheme
            type: http
            scheme: bearer
            bearerFormat: JWT # optional, arbitrary value for documentation purposes